requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
